+++ releaseNotes.md
@@ -21,6 +21,7 @@ edge / 2019-06-11
 - Breaking: getCellTypeDefinition is now getCellTypeDefinitionById
 - Breaking: getNodeTypeDefinitionByName is now getNodeTypeDefinitionByNodeTypeId
 - Breaking: getCompiledIndentation, getCompiledLine, getCompilerNode are now protected
+- Breaking: removed "constructors javascript" in GrammarLanguage
 - Infra: test coverage
 
 25.2.0 / 2019-05-30